Rename synctask package to task
